Refactor the code base
